Created independent Cheep record.
Cheep record has been added and TimeToCEST method has been added to Cheep, as it is only used in the record, resulting in better readability of the program as a whole.
