+++ UserInterface.cs
@@ -28,21 +28,6 @@ public static class UserInterface
     }
 }
 
-public record Cheep(string Author, string Message, long Timestamp)
-{
-    /// <summary>
-    /// Builder method for new Cheeps.
-    /// </summary>
-    /// <returns>A Cheep with the message content. Includes the current time, and env. username.</returns>
-    public static Cheep NewCheep(string message) 
-    {
-        return new Cheep(Environment.UserName, message, DateTimeOffset.UtcNow.ToUnixTimeSeconds());
-    }
 
-    override public string ToString()
-    {
-        return Author + " @ " + Program.TimecodeToCEST(Timestamp) + ": " + Message;
-    }
-}
